Extract more common methods with trivial implementations.
