+++ src/weak_weak_hash_map.rs
@@ -114,53 +114,7 @@ universal_hashless_members! {
 }
 
 impl<K: WeakKey, V: WeakElement, S: BuildHasher> WeakWeakHashMap<K, V, S> {
-    /// Removes all mappings whose keys have expired.
-    ///
-    /// *O*(*n*) time
-    pub fn remove_expired(&mut self) {
-        self.retain(|_, _| true);
-    }
-
-    /// Reserves room for additional elements.
-    ///
-    /// This method ensures that at least `additional_capacity` insertions
-    /// may be performed without reallocating.
-    ///
-    /// *O*(*n*) time
-    pub fn reserve(&mut self, additional_capacity: usize) {
-        self.try_reserve(additional_capacity)
-            .expect("Unable to reserve additional capacity");
-    }
-
-    /// Tries to reserve room for additional elements.
-    ///
-    /// If this method succeeds, then at least `additional_capacity` insertions
-    /// may be performed without reallocating further.
-    ///
-    /// *O*(*n*) time
-    pub fn try_reserve(
-        &mut self,
-        additional_capacity: usize,
-    ) -> Result<(), crate::TryReserveError> {
-        self.0.try_reserve(additional_capacity)
-    }
-
-    /// Shrinks the capacity to the minimum allowed to hold the current number of elements.
-    ///
-    /// *O*(*n*) time
-    pub fn shrink_to_fit(&mut self) {
-        self.0.shrink_to_fit();
-    }
-
-    /// Shrinks capacity to hold no fewer than `min_capacity` elements.
-    ///
-    /// May remove expired items if necessary.
-    /// Does nothing if the current capacity is already at `min_capacity` or below.
-    ///
-    /// *O*(*n*) time
-    pub fn shrink_to(&mut self, min_capacity: usize) {
-        self.0.shrink_to(min_capacity);
-    }
+    universal_key_independent_members! {"mappings"}
 
     /// Gets the requested entry.
     ///
